# apps/goa-app/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Project Template Goa Service
    description: A minimal Goa service setup for the project template repository.
    version: 0.0.1
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /auth/callback:
        post:
            tags:
                - oidc
            summary: complete oidc
            description: Exchange an authorization code for tokens.
            operationId: oidc#complete
            parameters:
                - name: CompleteRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/CompletePayload'
                    required:
                        - state
                        - code
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/OidcSession'
                "404":
                    description: Not Found response.
                    schema:
                        type: string
            schemes:
                - http
    /auth/start:
        post:
            tags:
                - oidc
            summary: start oidc
            description: Create a PKCE authorization request and return the redirect URL.
            operationId: oidc#start
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/OidcStart'
            schemes:
                - http
    /sessions/{state}:
        get:
            tags:
                - oidc
            summary: session oidc
            description: Retrieve the session associated with the provided state.
            operationId: oidc#session
            parameters:
                - name: state
                  in: path
                  description: State identifier
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/OidcSession'
                "404":
                    description: Not Found response.
                    schema:
                        type: string
            schemes:
                - http
    /status:
        get:
            tags:
                - status
            summary: show status
            description: Returns a simple service status message.
            operationId: status#show
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: string
            schemes:
                - http
definitions:
    CompletePayload:
        title: CompletePayload
        type: object
        properties:
            code:
                type: string
                description: Authorization code returned by the provider
                example: Molestiae consequatur nemo sint delectus at est.
            state:
                type: string
                description: State originally issued in the authorization request
                example: Qui numquam architecto quo nihil ratione voluptatum.
        example:
            code: Ex non non quae error.
            state: Quae iure recusandae aut doloribus.
        required:
            - state
            - code
    OidcSession:
        title: 'Mediatype identifier: application/vnd.oidc-session+json; view=default'
        type: object
        properties:
            access_token:
                type: string
                description: Opaque access token issued by the provider.
                example: Cupiditate mollitia officia reiciendis omnis perferendis.
            expires_in:
                type: integer
                description: Seconds until the access token expires.
                example: 4061380329348855495
                format: int64
            id_token:
                type: string
                description: ID token in JWT format.
                example: Et rem sed minima deserunt porro ab.
            scope:
                type: string
                description: Granted scopes.
                example: Et quo.
            user:
                $ref: '#/definitions/UserInfo'
        description: CompleteResponseBody result type (default view)
        example:
            access_token: In dolorum aut illum sapiente accusantium at.
            expires_in: 2484559224997405221
            id_token: Nihil aut repudiandae minus.
            scope: Dolorem doloremque.
            user:
                email: Voluptatibus molestiae placeat qui.
                name: Nemo corrupti amet nulla accusamus et.
                sub: Delectus velit nihil voluptatem at delectus.
        required:
            - access_token
            - id_token
            - expires_in
            - scope
            - user
    OidcStart:
        title: 'Mediatype identifier: application/vnd.oidc-start+json; view=default'
        type: object
        properties:
            auth_url:
                type: string
                description: Authorization endpoint URL with query parameters.
                example: Quam sit laborum vitae nesciunt odio nisi.
            state:
                type: string
                description: Opaque state value to correlate the login flow.
                example: Consequatur quo.
        description: StartResponseBody result type (default view)
        example:
            auth_url: Dolorem aliquam labore voluptatem ut.
            state: Fugit est necessitatibus totam laboriosam.
        required:
            - auth_url
            - state
    UserInfo:
        title: UserInfo
        type: object
        properties:
            email:
                type: string
                description: Email address
                example: Tenetur qui earum expedita.
            name:
                type: string
                description: Display name
                example: Dolorum ab maxime ut neque.
            sub:
                type: string
                description: Subject identifier
                example: Dolor at quam sequi reprehenderit ut.
        description: Subset of user attributes returned by the provider.
        example:
            email: Repudiandae inventore quae enim iste temporibus.
            name: Ea eum assumenda.
            sub: Et incidunt eveniet.
        required:
            - sub
            - email